Move org.embulk.deps.timestamp.DepsTimestampFormatter to org.embulk.spi.time.TimestampFormatterDelegate
