+++ embulk-deps/src/main/java/org/embulk/deps/timestamp/TimestampFormatterDelegateImpl.java
@@ -1,10 +1,11 @@
 package org.embulk.deps.timestamp;
 
 import java.time.Instant;
+import org.embulk.spi.time.TimestampFormatterDelegate;
 import org.embulk.util.timestamp.TimestampFormatter;
 
-public final class DepsTimestampFormatterImpl extends DepsTimestampFormatter {
-    public DepsTimestampFormatterImpl(final String pattern, final String defaultZone, final String defaultDate) {
+public final class TimestampFormatterDelegateImpl extends TimestampFormatterDelegate {
+    public TimestampFormatterDelegateImpl(final String pattern, final String defaultZone, final String defaultDate) {
         final TimestampFormatter.Builder builder = TimestampFormatter.builder(pattern, true);
         if (defaultZone != null) {
             builder.setDefaultZoneFromString(defaultZone);
